Fix missing sms or call.
The assumption of that the process which run the BroadcastReceiver is the same
process run the NotificationService might not always true, so simple put data
in a queue may not work if BroadcastReceiver.onReceive is called in another
process. This commit change the way of passing data, use Intent instead of
direct put the data into queue.